% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radviz.R
\name{add_color}
\alias{add_color}
\title{Add color to the charts}
\usage{
add_color(rv, color_accessor, palette = NULL)
}
\arguments{
\item{rv}{radviz object}

\item{color_accessor}{name of column/dimension to use for coloring}

\item{palette}{a vector of D3 colors that you wish to use (ordinal range that should
match the number of categoris in \code{color_accessor} (optional)}
}
\description{
Add color to the charts
}

