% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.r
\name{add_dimensions}
\alias{add_dimensions}
\title{Add the columns (dimensions) you wish to map in the radviz}
\usage{
add_dimensions(rv, dimensions)
}
\arguments{
\item{rv}{radviz object}

\item{dimesions}{character vector of column names}
}
\description{
Add the columns (dimensions) you wish to map in the radviz
}

