% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.r
\name{add_tooltip}
\alias{add_tooltip}
\title{Add a tooltip to a radviz widget}
\usage{
add_tooltip(rv, formatter = NULL)
}
\arguments{
\item{rv}{radviz object}

\item{formatter}{if \code{NULL} no tooltip will be added. Should be an atomic
character vector with javsascript code. It will be wrapped in a call to
\code{htmlwidgets::JS} (see examples)}
}
\description{
Add a tooltip to a radviz widget
}

