% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.radviz.R
\name{dim.radviz}
\alias{dim.radviz}
\title{Dimensions of a Radviz Object}
\usage{
\method{dim}{radviz}(x)
}
\arguments{
\item{x}{an object of class Radviz, as returned by \code{\link{do.radviz}}}
}
\description{
Retrieves the dimensions of a Radviz projection
}
\examples{
data(iris)
das <- c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')
S <- make.S(das)
rv <- do.radviz(iris,S)

dim(rv)
nrow(rv) # for free!

}
\author{
Yann Abraham
}

