% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.freeviz.R
\name{do.freeviz}
\alias{do.freeviz}
\title{Projects a Matrix or a Data Frame to a 2D space using Freeviz algorithm}
\usage{
do.freeviz(x, classes, attractG = 1, repelG = 1, law = 0,
  steps = 10, springs = NULL, print = TRUE)
}
\arguments{
\item{x}{Dataframe or matrix, with observations as rows and attributes as columns}

\item{classes}{Vector with class labels of the observations}

\item{attractG}{Number specifying the weight of the attractive forces}

\item{repelG}{Number specifying the weight of the repulsive forces}

\item{law}{Integer, specifying how forces change with distance: 0 = (inverse) linear, 1 = (inverse) square}

\item{steps}{Number of iterations of the algorithm before re-considering convergence criterion}

\item{springs}{Numeric matrix with initial anchor coordinates. When \code{NULL} (=default), springs are initialized by \code{\link{make.S}}}
}
\value{
An object of class radviz with the following slots:
         \itemize{
           \item \code{data} the original data (\code{x})
           \item \code{springs} the original \code{springs}
           \item \code{projected} the projection of \code{x} on \code{springs},
                   a matrix of 2D coordinates for every line in df
           \item \code{valid} a logical vector 
         }
}
\description{
Projects a Matrix or a Data Frame to a 2D space using Freeviz algorithm
}
\examples{
# the first example generates a simple Radviz object
data(iris)
fv <- do.freeviz(x = iris[,1:4], classes = iris$Species)
summary(fv)

}
\author{
Nicolas Sauwen
}
