% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.graphviz.R
\name{do.graphviz}
\alias{do.graphviz}
\title{Projects an attributes Matrix or a Data Frame with corresponding graph to a 2D space using Graphviz algorithm}
\usage{
do.graphviz(x, graph, attractG = 1, repelG = 1, law = 0,
  steps = 10, springs = NULL)
}
\arguments{
\item{x}{a data.frame or matrix to be projected, with column names matching row names in springs}

\item{graph}{\code{igraph} object}

\item{attractG}{Number specifying the weight of the attractive forces}

\item{repelG}{Number specifying the weight of the repulsive forces}

\item{law}{Integer, specifying how forces change with distance: 0 = (inverse) linear, 1 = (inverse) square}

\item{steps}{Number of iterations of the algorithm before re-considering convergence criterion}

\item{springs}{Numeric matrix with initial anchor coordinates. When \code{NULL} (=default), springs are initialized by \code{\link{make.S}}}
}
\value{
An object of class radviz with the following slots:
         \itemize{
           \item \code{data} the original data (\code{x})
           \item \code{springs} the computed \code{springs}
           \item \code{projected} the projection of \code{x} on \code{springs},
                   a matrix of 2D coordinates for every line in df
           \item \code{valid} a logical vector 
 		  \item \code{type} character string, indicating method used for computing the projection (in this case "graphviz")
           \item \code{graph} the original graph \code{igraph} object
         }
}
\description{
Projects an attributes Matrix or a Data Frame with corresponding graph to a 2D space using Graphviz algorithm
}
\author{
Nicolas Sauwen
}
