% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/do.L.R
\name{do.L}
\alias{do.L}
\title{Perform L-Normalization on a Vector}
\usage{
do.L(v, fun = range, na.rm = T)
}
\arguments{
\item{v}{a vector of values}

\item{fun}{a function that will return the minimum and maximum values to use to scale v;
defaults to \code{\link[base]{range}}}

\item{na.rm}{Logical: should NA be removed? defaults to \code{TRUE}}
}
\value{
A vector of values of the same lenght as x, scaled to the unit vector.
}
\description{
Standardizes all values in a vector to the unit vector ([0,1]) using local min and max
}
\details{
This is an alternative to performing a L normalization over the full matrix.
}
\examples{
data(iris)
mat <- iris[,c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')]
scaled <- apply(mat,2,do.L)
summary(scaled) # all values are between [0,1]
}
\author{
Yann Abraham
}

