% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.radviz.R
\name{identify.radviz}
\alias{identify.radviz}
\title{Identify a Point in a Radviz Projection}
\usage{
\method{identify}{radviz}(x, ...)
}
\arguments{
\item{x}{a radviz object as produced by \code{\link{do.radviz}}}

\item{...}{further arguments to be passed to or from other methods (not implemented)}
}
\value{
the original Radviz object
}
\description{
Use this function to get the index of a point in a Radviz projection
}
\details{
the do.hex function is not used anymore.
}
\examples{
data(iris)
das <- c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')
S <- make.S(das)
rv <- do.radviz(iris,S)
plot(rv,anchors.only=FALSE)

}
\author{
Yann Abraham
}
