% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.radviz.R
\name{is.radviz}
\alias{is.radviz}
\title{Radviz Object Summary, head, Print, Subset Methods}
\usage{
is.radviz(x)
}
\arguments{
\item{x}{an object of class Radviz, as returned by \code{\link{do.radviz}}}
}
\description{
Provides helper function to deal with Radviz objects
}
\examples{
data(iris)
das <- c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')
S <- make.S(das)
rv <- do.radviz(iris,S)

is.radviz(rv) # should be true
}
\author{
Yann Abraham
}

