% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.radviz.R
\name{print.radviz}
\alias{print.radviz}
\title{Print Radviz Object}
\usage{
\method{print}{radviz}(x, n = 6, ...)
}
\arguments{
\item{x}{an object of class Radviz, as returned by \code{\link{do.radviz}}}

\item{n}{the number of lines from each slots in the Radviz object to display (defaults to 6)}

\item{...}{further arguments to be passed to or from other methods}
}
\description{
Prints the first few lines of the data used to generate a Radviz projection
}
\examples{
data(iris)
das <- c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')
S <- make.S(das)
rv <- do.radviz(iris,S)

print(rv)

}
\author{
Yann Abraham
}
