% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radviz.R, R/do.density.R, R/do.hex.R,
%   R/identify.radviz.R
\name{Radviz-deprecated}
\alias{Radviz-deprecated}
\alias{do.density}
\alias{do.hex}
\alias{identify.radviz}
\title{Defunct Functions}
\usage{
do.density(x, ...)

do.hex(x, ...)

\method{identify}{radviz}(x, ...)
}
\description{
These functions are Defunct in this release of Radviz, they will be
marked as Defunct and removed in a future version.
}
\keyword{internal}
