% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radviz.R
\docType{package}
\name{Radviz}
\alias{Radviz}
\title{Radviz Projection of Multidimensional Data}
\description{
Radviz uses Dimensional Anchors and the spring paradigm to project a multidimensional space in 2D.
This allows for the quick visualization of large and complex datasets.
}
\examples{
data(iris)
das <- c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')
S <- make.S(das)
rv <- do.radviz(iris,S)
plot(rv,anchors.only=FALSE)

}
