% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.radviz.R
\name{subset.radviz}
\alias{subset.radviz}
\title{Title}
\usage{
\method{subset}{radviz}(x, i, ...)
}
\arguments{
\item{x}{a radviz object}

\item{i}{A logical or indices vector of the same length as the original data used to 
create the Radviz object, that is used to subset each slots}

\item{...}{further arguments to be passed to or from other methods}
}
\value{
a new Radviz object containing only rows specified in i.
         Any density or hexbin analysis is dropped
}
\description{
Title
}
\examples{
# create a radviz object
data(iris)
das <- c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')
S <- make.S(das)
rv <- do.radviz(iris,S)

# subset rv
srv <- subset(rv,iris$Species=='setosa')
summary(srv)
sum(iris$Species=='setosa') # 50 objects in srv corresponding to setosa values

}
\author{
Yann Abraham
}

