% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.radviz.R
\name{summary.radviz}
\alias{summary.radviz}
\title{Radviz Summary}
\usage{
\method{summary}{radviz}(object, n = 6, ...)
}
\arguments{
\item{object}{an object of class Radviz, as returned by \code{\link{do.radviz}}}

\item{n}{the number of lines from each slots in the Radviz object to display (defaults to 6)}

\item{...}{further arguments to be passed to or from other methods}
}
\description{
Provides a summary for Radviz objects
}
\examples{
data(iris)
das <- c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')
S <- make.S(das)
rv <- do.radviz(iris,S)

summary(rv)

}
\author{
Yann Abraham
}

