% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text.radviz.R
\name{text.radviz}
\alias{text.radviz}
\title{Text annotations for for the Radviz Plots}
\usage{
\method{text}{radviz}(x, ..., main = NULL, label.color = "orangered4",
  label.size = 1, labels = NULL, adj = NULL, pos = NULL, offset = 0.5,
  vfont = NULL, cex = 1, col = NULL, font = NULL, add = TRUE)
}
\arguments{
\item{x}{a radviz object as produced by do.radviz}

\item{...}{further arguments to be passed to or from other methods}

\item{main}{[Optional] a title to the graph, displayed on top if add is \code{TRUE}}

\item{label.color}{[Optional] The color of the Dimensional Anchors if add is
\code{TRUE} (defaults to orangered4)}

\item{label.size}{[Optional] numeric character expansion factor for Dimensional Anchor
labels if add is \code{TRUE}; multiplied by \code{par("cex")} yields the final
character size. NULL and NA are equivalent to 1.0}

\item{labels}{a character vector specifying the text to be written. An attempt is made
to coerce other language objects (names and calls) to characters using
  \code{\link[grDevices]{as.graphicsAnnot}}. If labels is longer than x and y,
  the coordinates are recycled to the length of labels.}

\item{adj}{one or two values in [0, 1] which specify the x (and optionally y) adjustment
of the labels. On most devices values outside that interval will also work.}

\item{pos}{a position specifier for the text. If specified this overrides any adj value given.
Values of 1, 2, 3 and 4, respectively indicate positions below, to the left of, above
and to the right of the specified coordinates.}

\item{offset}{when \code{pos} is specified, this value gives the offset of the label from the
specified coordinate in fractions of a character width.}

\item{vfont}{\code{NULL} for the current font family, or a character vector of length 2 for
Hershey vector fonts. The first element of the vector selects a typeface and the second
element selects a style. Ignored if labels is an expression.}

\item{cex}{numeric character expansion factor; multiplied by \code{\link{par}("cex")} yields
the final character size. NULL and NA are equivalent to 1.0.}

\item{col,font}{the color and (if vfont = NULL) font to be used, possibly vectors.
These default to the values of the global \link{graphical parameters} in \link{par}().}

\item{add}{Logical: if add is \code{TRUE} then only the projected points are plotted}
}
\description{
Text draws the strings given in the vector labels at the coordinates given by the radviz projection
}
\examples{
data(iris)
das <- c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')
S <- make.S(das)
rv <- do.radviz(iris,S)
plot(rv,point.shape=1,point.color='grey')
med.iris <- split(iris,iris$Species)
med.iris <- lapply(med.iris,function(df) {
   spc <- unique(df$Species)
   df <-df[,names(df)!='Species']
   df <- apply(df,2,median)
   df <- data.frame(t(df))
   df$Species <- spc
   return(df)
 }
)
med.iris <- do.call('rbind',med.iris)
med.rv <- do.radviz(med.iris,S)
text(med.rv,labels=med.iris$Species,col=c('red','green','blue')[as.integer(med.iris$Species)])
}
\author{
Yann Abraham
}

