% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_weightedIdx.R
\name{DB_weightedIdx}
\alias{DB_weightedIdx}
\title{Computation of weighted version of the Davies-Bouldin index. This index serves as a measure of clustering quality of a 2D projection result with known class labels}
\usage{
DB_weightedIdx(x)
}
\arguments{
\item{x}{an object of class Radviz, as returned by \code{\link{do.freeviz}}. Note that the DB weighted index can only be computed for Radviz objects with slot \code{type} = "freeviz"}
}
\value{
weighted DB index value
}
\description{
Computation of weighted version of the Davies-Bouldin index. This index serves as a measure of clustering quality of a 2D projection result with known class labels
}
\author{
Nicolas Sauwen
}
