% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_weightedIdx.R
\name{DB_weightedIdx}
\alias{DB_weightedIdx}
\title{Computation of weighted version of the Davies-Bouldin index. This index serves as a measure of clustering quality of a 2D projection result with known class labels}
\usage{
DB_weightedIdx(x, className = NULL)
}
\arguments{
\item{x}{an object of class Radviz, as returned by \code{\link{do.radviz}}}

\item{className}{the name of the class column to use}
}
\value{
weighted DB index value
}
\description{
Computation of weighted version of the Davies-Bouldin index. This index serves as a measure of clustering quality of a 2D projection result with known class labels
}
\details{
If \code{className} is left \code{NULL} (the default) the function expects a single extra column on top of 
the data columns (used to define springs) and the standard \code{Radviz} columns.
}
\author{
Nicolas Sauwen
}
