% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.density.R
\name{do.density}
\alias{do.density}
\title{Computes 2D density for contour plots of Radviz objects}
\usage{
do.density(x, ...)
}
\arguments{
\item{x}{a radviz object as produced by do.radviz}

\item{...}{further arguments to be passed to or from other methods (not implemented)}
}
\value{
the original Radviz object
}
\description{
Computes 2D density estimate of projected data for a Radivz object,
using the \code{\link[MASS]{kde2d}} function from the \pkg{MASS} package
}
\details{
the do.hex function is not required anymore, use \code{\link{contour.radviz}} directly.
}
\examples{
data(iris)
das <- c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')
S <- make.S(das)
rv <- do.radviz(iris,S)
contour(rv)

}
\seealso{
\link{contour.radviz} for plotting
}
\author{
Yann Abraham
}
