% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.hex.R
\name{do.hex}
\alias{do.hex}
\title{Computes 2D hexagonal bins from a Radviz projection}
\usage{
do.hex(x, ...)
}
\arguments{
\item{x}{a radviz object as produced by do.radviz}

\item{...}{further arguments to be passed to or from other methods (not implemented)}
}
\value{
the original Radviz object
}
\description{
Computes 2D density using hexagon binning of projected data for a Radivz object,
using the \code{\link[hexbin]{hexbin}} function from the \pkg{hexbin} package
}
\details{
the do.hex function is not required anymore, use \code{\link{hexplot}} directly.
}
\examples{
data(iris)
das <- c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')
S <- make.S(das)
rv <- do.radviz(iris,S)
hexplot(rv)

}
\seealso{
\link{hexplot} for plotting
}
\author{
Yann Abraham
}
