% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/do.radviz.R
\name{do.radviz}
\alias{do.radviz}
\alias{do.radviz.default}
\title{Projects a Matrix or a Data Frame to a 2D space defined by Dimensional Anchors}
\usage{
do.radviz(x, springs)
}
\arguments{
\item{x}{a placeholder for defining the S3 method; currently, expects a data.frame
or matrix to be projected, with column names matching row names in springs}

\item{springs}{A matrix of 2D dimensional anchor coordinates, as returned by \code{\link{make.S}}}
}
\value{
An object of class radviz with the following slots:
         \itemize{
           \item \code{data} the original data (\code{x})
           \item \code{springs} the original \code{springs}
           \item \code{projected} the projection of \code{x} on \code{springs},
                   a matrix of 2D coordinates for every line in df
         }
}
\description{
do.radviz will return a projection of a multidimensional dataset onto a 2D space
defined by dimensional anchors that have been projected on the unit circle using
\code{\link{make.S}}
}
\details{
The function expects that at least some of the column names in df will be matched
           by row names in springs
}
\examples{
data(iris)
das <- c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')
S <- make.S(das)
rv <- do.radviz(iris,S)
summary(rv)
}
\author{
Yann Abraham
}

