% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.radviz.R
\name{head.radviz}
\alias{head.radviz}
\title{Radviz Object Summary, head, Print, Subset Methods}
\usage{
\method{head}{radviz}(x, n = 6, ...)
}
\arguments{
\item{x}{an object of class Radviz, as returned by \code{\link{do.radviz}}}

\item{n}{the number of lines from each slots in the Radviz object to display (defaults to 6)}

\item{...}{further arguments to be passed to or from other methods}
}
\description{
Provides helper function to deal with Radviz objects
}
\details{
\code{head.radviz} shows the first \code{n} lines of the radviz \code{data} object.
}
\examples{
data(iris)
das <- c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')
S <- make.S(das)
rv <- do.radviz(iris,S)

head(rv)

}
\author{
Yann Abraham
}

