% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.radviz.R
\name{is.radviz}
\alias{is.radviz}
\title{Test if the object is a Radviz object}
\usage{
is.radviz(x)
}
\arguments{
\item{x}{an object of class Radviz, as returned by \code{\link{do.radviz}}}
}
\description{
The function will return \code{TRUE} if the object is a Radviz object
}
\examples{
data(iris)
das <- c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')
S <- make.S(das)
rv <- do.radviz(iris,S)
is.radviz(rv) # should be true

}
\author{
Yann Abraham
}
