% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radviz.R
\name{radviz}
\alias{radviz}
\title{Make a radviz}
\usage{
radviz(data, draw_links = TRUE, zoom_factor = 1, dot_radius = 4,
  use_repulsion = FALSE, diameter = 400, margin = 50, width = NULL,
  height = NULL)
}
\arguments{
\item{data}{\code{data.frame} to use}

\item{draw_links}{whether to draw links or not}

\item{zoom_factor}{gets the dimension nodes on a bigger radius than the panel, to pull nodes away from the center}

\item{dot_radius}{radius of each dot}

\item{use_repulsion}{special repulsion effect to spread the nodes from each others so they can be better selected}

\item{diameter}{size of the circle (should be <= \code{width} if that is set to pixels)}

\item{margin}{margin around the circular panel}

\item{width, height}{height & width of the widget}
}
\description{
Make a radviz
}
\examples{
mtcars_1 <- dplyr::add_rownames(mtcars, var="car")

radviz(mtcars_1, diameter=600, margin=100,
       use_repulsion=TRUE, draw_links=TRUE,
       width=600, height=500) \%>\%
  add_dimensions(c("mpg", "cyl", "disp", "hp", "drat", "wt",
                   "qsec", "vs", "am", "gear", "carb")) \%>\%
  add_color("cyl") \%>\%
  add_tooltip("function(d) { return d.car; }")
}

