% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.radviz.R
\name{summary.radviz}
\alias{head.radviz}
\alias{print.radviz}
\alias{subset.radviz}
\alias{summary.radviz}
\title{Radviz Object Summary, Head, Print, Subset Methods}
\usage{
summary(x,n=8)
head(x,n=6)
print(x,n=6)
subset(x,i)
}
\arguments{
\item{x}{An object of class Radviz, as returned by \code{link{do.radviz}}}

\item{n}{The number of lines from each slots in the Radviz object to display (defaults to 6)}

\item{i}{A logical or indices vector of the same length as the original data used to
         create the Radviz object, that is used to subset each slots}
}
\value{
For subset.radviz, a new Radviz object containing only rows specified in i.
         Any density or hexbin analysis is dropped
}
\description{
Provides helper function to deal with Radviz objects
}
\examples{
data(iris)
das <- c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')
S <- make.S(das)
rv <- do.radviz(iris,S)
summary(rv)
head(rv,5)
print(rv)
# create a subset
srv <- subset(rv,iris$Species=='setosa')
summary(srv)
sum(iris$Species=='setosa') # 50 objects in srv corresponding to setosa values
}
\author{
Yann Abraham
}

